Fix item deletion by using parseFloat instead of parseInt for IDs
- Item IDs are floats (Date.now() + Math.random()), not integers
- Was failing to match and delete items properly

+++ extension/content.js
@@ -582,7 +582,7 @@ function renderAgendaItems(overlay) {
   deleteButtons.forEach((btn) => {
     btn.addEventListener('click', (e) => {
       e.stopPropagation();
-      const itemId = parseInt(btn.dataset.itemId);
+      const itemId = parseFloat(btn.dataset.itemId);
       deleteAgendaItemFromOverlay(itemId);
     });
   });
